Organize CLI scripts and repo metadata within $HOME/.dotfiles subdirectory
In an effort to keep the $HOME directory clean, CLI scripts and
metadata (e.g. the README.md file) will be stored within the
`.dotfiles` directory within the user home folder.

We can be smart about it and link the README.md to the git working
directory so that it continues to display on the repository webpage.