Add cleaned-up ZSH configuration file
I spent some initial efforts cleaning up and documenting my ZSH file.
It was nice to go through and clean out some gunk that has accumulated.

Everything here is decently modularized. Plugins have a home in
~/.dotfiles/plugins as git submodules, but a lot of dependencies
referenced in the .zshrc are still unindexed as plugins. Eventually I'd
like to track these so that an environment restore can be replicated.

The two instances of auth tokens exported as environment variables have
been updated to pull from the macOS keychain. This feels like a good
pattern; if not, perhaps more so if simplified with a function.