Install language server to Neovim
These changes use Nvim's support for the Language Server Protocol (LSP)
to provide commands for navigating, modifying and inspecting code.

The lspconfig package sets sensible defaults. Custom configuration is
declared in that package's config file, since it's aptly named.

Inside of this file is a configuration for nvim-compe, which provides
setup and bindings for auto-completion.

The lspinstall package adds the `:LspInstall <language>` command, which
handles installing the client-side processes necessary for language
support.